fix view not updated when decoration closure changes
Before, CalendarView only reloaded decorations when the date components changed but not if the decorations themselves changed while the date components stayed the same. To fix this, the modifier `CalendarView.decorating` now takes an `Equatable` parameter that ensures that the view is updated (i.e. `updateUIView` is called). Furthermore, each view update now reloads the decorations of all currently visible date components instead of only the ones that changed. This makes sure that the decoration closure gets called when changes occur (either to the date components or their decorations.)

In addition, decorations are now always reloaded for the visible month to ensure that decorations are updated if only the decorations changed but not the date components.